Create short links for battles
This proposed update creates short links and prints the short links. The $domainname needs to be declared on line 311 and the $customdir needs to be manually created (and optionally changed).  This is for easier sharing of battles so the links do not seem quite as spammy.

Feel free to view the live example at https://prts.cr/calc/ .  This example is planned to be removed after PR is accepted/rejected.

This update is intended to be used with a cronjob similar to `0 0 * * * find /public_html/calc/b -depth -type f -atime +10 -delete` which will delete the redirect webpages (in the `b` directory) after they have not been accessed for 10 days, and will run on the first of the month. Keep in mind, everytime the redirect webpage is loaded, it will create a new shortlink, so expect quite a few redirects per actual battle. 

I am seeing 780B per webpage which is nominal if they are deleted after 10 (or more days) but will possibly be burdensome if they are never deleted.

It is currently unknown why the bottom link has moved to the top.